Update sidmovement.cs
This is just a backup before I go crazy with adding the Multi-tap interaction for the dive-roll, which will take me a while to figure out. Once I figure out double presses then adding dive-rolls and other movements will be easy. This script has a good amount of bugs but overall works in Unity.